# src/swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "API  EBMS",
    "description": "Essa api tem como objetivo ajudar as doações de sangue.",
    "contact": {
      "email": "saveambiente@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://ebms-api.onrender.com"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Cria um novo administrador.",
        "operationId": "createAdmin",
        "tags": ["Administração"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cria uma conta de administrador.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas ou ausentes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/agendamentos/": {
      "post": {
        "summary": "Cadastra agendamento.",
        "operationId": "criarAgendamento",
        "tags": ["Agendamento"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/agendamento"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cadastado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cheque as informações.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Consulta agendamento.",
        "description": "Consulta todos os agendamentos.",
        "operationId": "listarAgendamentos",
        "tags": ["Agendamento"],
        "responses": {
          "200": {
            "description": "Consulta com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/agendamentos/:id": {
      "put": {
        "summary": "Atualiza dados do agendamento buscado pelo id.",
        "operationId": "atualizarAgendamento",
        "tags": ["Agendamento"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Id para busca.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/agendamentoUp"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cheque o id.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deleta validade.",
        "description": "Deleta validade usando ID.",
        "operationId": "deletarAgendamento",
        "tags": ["Agendamento"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Id para busca.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "agendamento":{
        "type": "object",
        "properties": {
          "usuario_id": {
            "type": "string",
            "description": "Userid"
          }, 
          "data_agendamento":{
            "type": "string",
            "description": "Data de doação"
          }, 
          "local": {
            "type": "string",
            "description": "local da doação"
          }, 
          "status": {
            "type": "string",
            "description": "Estado do agendamento"
          }
        }
      },
      "agendamentoUp":{
        "type": "object",
        "properties": {
          "data_agendamento":{
            "type": "string",
            "description": "Data de doação"
          }, 
          "local": {
            "type": "string",
            "description": "local da doação"
          }, 
          "status": {
            "type": "string",
            "description": "Estado do agendamento"
          }
        }
      }
    }
  }
}